Optimised putmem. Doesn't need seperate hex function anymore!

+++ ft_hex.c
@@ -6,13 +6,13 @@
 /*   By: mialbert <mialbert@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/05 02:10:29 by mialbert          #+#    #+#             */
-/*   Updated: 2021/12/16 21:52:29 by mialbert         ###   ########.fr       */
+/*   Updated: 2021/12/18 22:44:03 by mialbert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int32_t	ft_hex(uint32_t nbr, const char *hexbase)
+int32_t	ft_hex(uint64_t nbr, const char *hexbase)
 {
 	int32_t	i;
